make sure that test will be marked as finished in tes explorer by preventing exception

+++ src/JSNLog.Tests/IntegrationTests/JsTestsContext.cs
@@ -69,7 +69,11 @@ namespace JSNLog.Tests.IntegrationTests
             // Close the browser if there is no error. Otherwise leave open.
             if (!ErrorOnPage())
             {
-                _serverProcess.Kill();
+                if (!_serverProcess.HasExited)
+                {
+                    _serverProcess.Kill();
+                }
+
                 Driver.Quit();
             }
         }
